vfs split
