+++ crates/vfs/src/lib.rs
@@ -1,1259 +1,16 @@
 //! CAS-backed virtual filesystem models and snapshot helpers.
 //!
 //! This crate owns deterministic VFS path and manifest structures plus helpers
-//! that write immutable snapshot manifests into an injected [`BlobStore`].
+//! that write immutable snapshot manifests into an injected `BlobStore`.
 //! Host filesystem access, materialization, and process execution live outside
 //! this crate.
 
-use async_trait::async_trait;
-use engine::{
-    BlobRef, SessionId, StringIdError,
-    storage::{BlobStore, BlobStoreError},
-    validate_general_string_id,
-};
-use serde::{Deserialize, Deserializer, Serialize, Serializer, de};
-use std::{
-    collections::{BTreeMap, BTreeSet},
-    fmt,
-    str::FromStr,
-};
-use thiserror::Error;
-
-pub const VFS_SNAPSHOT_SCHEMA_VERSION: &str = "forge.vfs.snapshot.v1";
-
-macro_rules! vfs_string_id {
-    ($name:ident) => {
-        #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-        pub struct $name(String);
-
-        impl $name {
-            pub fn new(value: impl Into<String>) -> Self {
-                let value = value.into();
-                Self::try_new(value)
-                    .unwrap_or_else(|error| panic!("invalid {}: {error}", stringify!($name)))
-            }
-
-            pub fn try_new(value: impl Into<String>) -> Result<Self, StringIdError> {
-                let value = value.into();
-                validate_general_string_id(stringify!($name), &value)?;
-                Ok(Self(value))
-            }
-
-            pub fn parse(value: impl Into<String>) -> Result<Self, StringIdError> {
-                Self::try_new(value)
-            }
-
-            pub fn as_str(&self) -> &str {
-                &self.0
-            }
-        }
-
-        impl TryFrom<String> for $name {
-            type Error = StringIdError;
-
-            fn try_from(value: String) -> Result<Self, Self::Error> {
-                Self::try_new(value)
-            }
-        }
-
-        impl TryFrom<&str> for $name {
-            type Error = StringIdError;
-
-            fn try_from(value: &str) -> Result<Self, Self::Error> {
-                Self::try_new(value)
-            }
-        }
-
-        impl FromStr for $name {
-            type Err = StringIdError;
-
-            fn from_str(value: &str) -> Result<Self, Self::Err> {
-                Self::try_new(value)
-            }
-        }
-
-        impl Serialize for $name {
-            fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-            where
-                S: Serializer,
-            {
-                serializer.serialize_str(&self.0)
-            }
-        }
-
-        impl<'de> Deserialize<'de> for $name {
-            fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-            where
-                D: Deserializer<'de>,
-            {
-                let value = String::deserialize(deserializer)?;
-                Self::try_new(value).map_err(de::Error::custom)
-            }
-        }
-
-        impl fmt::Display for $name {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                f.write_str(&self.0)
-            }
-        }
-    };
-}
-
-vfs_string_id!(VfsWorkspaceId);
-
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-#[serde(transparent)]
-pub struct VfsPath(String);
-
-impl VfsPath {
-    pub fn parse(value: impl AsRef<str>) -> Result<Self, VfsPathError> {
-        let value = value.as_ref();
-        if value.is_empty() {
-            return Err(VfsPathError::Empty);
-        }
-        if value.as_bytes().contains(&0) {
-            return Err(VfsPathError::ContainsNul {
-                path: value.to_owned(),
-            });
-        }
-        if value == "/" {
-            return Ok(Self("/".to_owned()));
-        }
-
-        let trimmed = value.strip_prefix('/').unwrap_or(value);
-        if trimmed.is_empty() {
-            return Err(VfsPathError::InvalidComponent {
-                path: value.to_owned(),
-                component: String::new(),
-            });
-        }
-
-        let mut components = Vec::new();
-        for component in trimmed.split('/') {
-            if component.is_empty() || component == "." || component == ".." {
-                return Err(VfsPathError::InvalidComponent {
-                    path: value.to_owned(),
-                    component: component.to_owned(),
-                });
-            }
-            components.push(component);
-        }
-
-        Ok(Self(format!("/{}", components.join("/"))))
-    }
-
-    pub fn root() -> Self {
-        Self("/".to_owned())
-    }
-
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-
-    pub fn is_root(&self) -> bool {
-        self.0 == "/"
-    }
-
-    pub fn depth(&self) -> usize {
-        self.components().len()
-    }
-
-    pub fn components(&self) -> Vec<&str> {
-        if self.is_root() {
-            Vec::new()
-        } else {
-            self.0.trim_start_matches('/').split('/').collect()
-        }
-    }
-}
-
-impl fmt::Display for VfsPath {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-impl TryFrom<&str> for VfsPath {
-    type Error = VfsPathError;
-
-    fn try_from(value: &str) -> Result<Self, Self::Error> {
-        Self::parse(value)
-    }
-}
-
-impl TryFrom<String> for VfsPath {
-    type Error = VfsPathError;
-
-    fn try_from(value: String) -> Result<Self, Self::Error> {
-        Self::parse(value)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Error)]
-pub enum VfsPathError {
-    #[error("vfs path must not be empty")]
-    Empty,
-
-    #[error("vfs path contains a NUL byte: {path}")]
-    ContainsNul { path: String },
-
-    #[error("vfs path has an invalid component '{component}': {path}")]
-    InvalidComponent { path: String, component: String },
-}
-
-#[derive(Debug, Error)]
-pub enum VfsError {
-    #[error(transparent)]
-    InvalidPath(#[from] VfsPathError),
-
-    #[error("vfs file path cannot be the root path")]
-    RootFile,
-
-    #[error("duplicate vfs path: {path}")]
-    DuplicatePath { path: VfsPath },
-
-    #[error("vfs path conflicts with an existing {existing}: {path}")]
-    PathConflict {
-        path: VfsPath,
-        existing: &'static str,
-    },
-
-    #[error("vfs path not found: {path}")]
-    NotFound { path: VfsPath },
-
-    #[error("vfs path is not a file: {path}")]
-    NotAFile { path: VfsPath },
-
-    #[error("vfs path is not a directory: {path}")]
-    NotADirectory { path: VfsPath },
-
-    #[error("vfs snapshot limit exceeded for {limit}: value {value} is greater than max {max}")]
-    LimitExceeded {
-        limit: &'static str,
-        value: u64,
-        max: u64,
-    },
-
-    #[error("invalid vfs snapshot manifest: {message}")]
-    InvalidManifest { message: String },
-
-    #[error("failed to encode vfs snapshot manifest: {source}")]
-    EncodeManifest {
-        #[source]
-        source: serde_json::Error,
-    },
-
-    #[error("failed to decode vfs snapshot manifest: {source}")]
-    DecodeManifest {
-        #[source]
-        source: serde_json::Error,
-    },
-
-    #[error(transparent)]
-    BlobStore(#[from] BlobStoreError),
-}
-
-#[derive(Debug, Error)]
-pub enum VfsCatalogError {
-    #[error("vfs catalog {kind} already exists: {id}")]
-    AlreadyExists { kind: &'static str, id: String },
-
-    #[error("vfs catalog {kind} not found: {id}")]
-    NotFound { kind: &'static str, id: String },
-
-    #[error(
-        "vfs workspace revision conflict for {workspace_id}: expected {expected_revision}, actual {actual_revision}"
-    )]
-    RevisionConflict {
-        workspace_id: VfsWorkspaceId,
-        expected_revision: u64,
-        actual_revision: u64,
-    },
-
-    #[error("invalid vfs catalog request: {message}")]
-    InvalidInput { message: String },
-
-    #[error("vfs catalog store failure: {message}")]
-    Store { message: String },
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct VfsSnapshotManifest {
-    pub schema_version: String,
-    pub root: VfsDirectory,
-    pub totals: VfsTotals,
-}
-
-impl VfsSnapshotManifest {
-    pub fn empty() -> Self {
-        Self {
-            schema_version: VFS_SNAPSHOT_SCHEMA_VERSION.to_owned(),
-            root: VfsDirectory::default(),
-            totals: VfsTotals::default(),
-        }
-    }
-
-    pub fn validate(&self) -> Result<(), VfsError> {
-        if self.schema_version != VFS_SNAPSHOT_SCHEMA_VERSION {
-            return Err(VfsError::InvalidManifest {
-                message: format!("unsupported schema version '{}'", self.schema_version),
-            });
-        }
-        let mut totals = VfsTotals::default();
-        validate_directory(&self.root, &mut totals)?;
-        if totals != self.totals {
-            return Err(VfsError::InvalidManifest {
-                message: format!(
-                    "manifest totals do not match tree contents: expected {:?}, computed {:?}",
-                    self.totals, totals
-                ),
-            });
-        }
-        Ok(())
-    }
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct VfsDirectory {
-    #[serde(default)]
-    pub entries: BTreeMap<String, VfsEntry>,
-}
-
-impl VfsDirectory {
-    pub fn entry(&self, name: &str) -> Option<&VfsEntry> {
-        self.entries.get(name)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum VfsEntry {
-    File(VfsFile),
-    Directory(VfsDirectory),
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct VfsFile {
-    pub blob_ref: BlobRef,
-    pub size_bytes: u64,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub media_type: Option<String>,
-    pub executable: bool,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct VfsTotals {
-    pub files: u64,
-    pub bytes: u64,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct VfsSnapshotRecord {
-    pub snapshot_ref: BlobRef,
-    pub source: VfsSnapshotSource,
-    pub display_name: Option<String>,
-    pub created_at_ms: i64,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct VfsSnapshotSource {
-    pub kind: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub subject: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub metadata_ref: Option<BlobRef>,
-}
-
-impl VfsSnapshotSource {
-    pub fn new(kind: impl Into<String>) -> Self {
-        Self {
-            kind: kind.into(),
-            subject: None,
-            metadata_ref: None,
-        }
-    }
-
-    pub fn unknown() -> Self {
-        Self::new("unknown")
-    }
-
-    pub fn with_subject(mut self, subject: impl Into<String>) -> Self {
-        self.subject = Some(subject.into());
-        self
-    }
-
-    pub fn with_metadata_ref(mut self, metadata_ref: BlobRef) -> Self {
-        self.metadata_ref = Some(metadata_ref);
-        self
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct VfsWorkspaceRecord {
-    pub workspace_id: VfsWorkspaceId,
-    pub base_snapshot_ref: Option<BlobRef>,
-    pub head_snapshot_ref: BlobRef,
-    pub revision: u64,
-    pub created_at_ms: i64,
-    pub updated_at_ms: i64,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CreateVfsWorkspaceRecord {
-    pub workspace_id: VfsWorkspaceId,
-    pub base_snapshot_ref: Option<BlobRef>,
-    pub head_snapshot_ref: BlobRef,
-    pub created_at_ms: i64,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CompareAndSetVfsWorkspaceHead {
-    pub workspace_id: VfsWorkspaceId,
-    pub expected_revision: u64,
-    pub new_head_snapshot_ref: BlobRef,
-    pub updated_at_ms: i64,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct VfsMountRecord {
-    pub session_id: SessionId,
-    pub mount_path: VfsPath,
-    pub source: VfsMountSource,
-    pub access: VfsMountAccess,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum VfsMountSource {
-    Snapshot { snapshot_ref: BlobRef },
-    Workspace { workspace_id: VfsWorkspaceId },
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum VfsMountAccess {
-    ReadOnly,
-    ReadWrite,
-}
-
-impl VfsMountAccess {
-    pub const fn is_writable(self) -> bool {
-        matches!(self, Self::ReadWrite)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct VfsMountTable {
-    pub session_id: SessionId,
-    pub mounts: Vec<VfsMountRecord>,
-}
-
-#[async_trait]
-pub trait VfsSnapshotCatalog: Send + Sync {
-    async fn record_snapshot(&self, record: VfsSnapshotRecord) -> Result<(), VfsCatalogError>;
-
-    async fn read_snapshot(
-        &self,
-        snapshot_ref: &BlobRef,
-    ) -> Result<VfsSnapshotRecord, VfsCatalogError>;
-}
-
-#[async_trait]
-pub trait VfsWorkspaceStore: Send + Sync {
-    async fn create_workspace(
-        &self,
-        record: CreateVfsWorkspaceRecord,
-    ) -> Result<VfsWorkspaceRecord, VfsCatalogError>;
-
-    async fn read_workspace(
-        &self,
-        workspace_id: &VfsWorkspaceId,
-    ) -> Result<VfsWorkspaceRecord, VfsCatalogError>;
-
-    async fn compare_and_set_head(
-        &self,
-        request: CompareAndSetVfsWorkspaceHead,
-    ) -> Result<VfsWorkspaceRecord, VfsCatalogError>;
-}
-
-#[async_trait]
-pub trait VfsMountStore: Send + Sync {
-    async fn put_mount(&self, record: VfsMountRecord) -> Result<(), VfsCatalogError>;
-
-    async fn list_mounts(
-        &self,
-        session_id: &SessionId,
-    ) -> Result<Vec<VfsMountRecord>, VfsCatalogError>;
-
-    async fn remove_mount(
-        &self,
-        session_id: &SessionId,
-        mount_path: &VfsPath,
-    ) -> Result<(), VfsCatalogError>;
-}
-
-pub trait VfsCatalogStore: VfsSnapshotCatalog + VfsWorkspaceStore + VfsMountStore {}
-
-impl<T> VfsCatalogStore for T where T: VfsSnapshotCatalog + VfsWorkspaceStore + VfsMountStore {}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct InlineFile {
-    pub path: VfsPath,
-    pub bytes: Vec<u8>,
-    pub media_type: Option<String>,
-    pub executable: bool,
-}
-
-impl InlineFile {
-    pub fn new(path: impl AsRef<str>, bytes: impl Into<Vec<u8>>) -> Result<Self, VfsError> {
-        Ok(Self {
-            path: VfsPath::parse(path)?,
-            bytes: bytes.into(),
-            media_type: None,
-            executable: false,
-        })
-    }
-
-    pub fn with_media_type(mut self, media_type: impl Into<String>) -> Self {
-        self.media_type = Some(media_type.into());
-        self
-    }
-
-    pub fn executable(mut self, executable: bool) -> Self {
-        self.executable = executable;
-        self
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct VfsSnapshotLimits {
-    pub max_files: u64,
-    pub max_total_bytes: u64,
-    pub max_single_file_bytes: u64,
-    pub max_depth: usize,
-}
-
-impl VfsSnapshotLimits {
-    pub const fn new(
-        max_files: u64,
-        max_total_bytes: u64,
-        max_single_file_bytes: u64,
-        max_depth: usize,
-    ) -> Self {
-        Self {
-            max_files,
-            max_total_bytes,
-            max_single_file_bytes,
-            max_depth,
-        }
-    }
-}
-
-impl Default for VfsSnapshotLimits {
-    fn default() -> Self {
-        Self {
-            max_files: 10_000,
-            max_total_bytes: 512 * 1024 * 1024,
-            max_single_file_bytes: 128 * 1024 * 1024,
-            max_depth: 64,
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct CreateInlineSnapshotRequest {
-    pub files: Vec<InlineFile>,
-    pub limits: VfsSnapshotLimits,
-}
-
-impl CreateInlineSnapshotRequest {
-    pub fn new(files: Vec<InlineFile>) -> Self {
-        Self {
-            files,
-            limits: VfsSnapshotLimits::default(),
-        }
-    }
-
-    pub fn with_limits(mut self, limits: VfsSnapshotLimits) -> Self {
-        self.limits = limits;
-        self
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct CreateVfsSnapshotResult {
-    pub snapshot_ref: BlobRef,
-    pub manifest: VfsSnapshotManifest,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum VfsNode<'a> {
-    File(&'a VfsFile),
-    Directory(&'a VfsDirectory),
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct VfsPathMetadata {
-    pub is_directory: bool,
-    pub is_file: bool,
-    pub size_bytes: Option<u64>,
-    pub media_type: Option<String>,
-    pub executable: bool,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct VfsDirectoryListingEntry {
-    pub file_name: String,
-    pub is_directory: bool,
-    pub is_file: bool,
-    pub size_bytes: Option<u64>,
-}
-
-pub async fn create_inline_snapshot(
-    blobs: &(impl BlobStore + ?Sized),
-    request: CreateInlineSnapshotRequest,
-) -> Result<CreateVfsSnapshotResult, VfsError> {
-    validate_inline_files(&request.files, request.limits)?;
-
-    let mut staged_files = Vec::with_capacity(request.files.len());
-    let mut file_bytes = Vec::with_capacity(request.files.len());
-    for file in request.files {
-        staged_files.push((
-            file.path,
-            file.bytes.len() as u64,
-            file.media_type,
-            file.executable,
-        ));
-        file_bytes.push(file.bytes);
-    }
-
-    let blob_refs = blobs.put_many(file_bytes).await?;
-    let mut manifest = VfsSnapshotManifest::empty();
-    for ((path, size_bytes, media_type, executable), blob_ref) in
-        staged_files.into_iter().zip(blob_refs)
-    {
-        let vfs_file = VfsFile {
-            blob_ref,
-            size_bytes,
-            media_type,
-            executable,
-        };
-        insert_file(&mut manifest.root, &path, vfs_file)?;
-        manifest.totals.files += 1;
-        manifest.totals.bytes += size_bytes;
-    }
-
-    let manifest_bytes = encode_snapshot_manifest(&manifest)?;
-    let snapshot_ref = blobs.put_bytes(manifest_bytes).await?;
-    Ok(CreateVfsSnapshotResult {
-        snapshot_ref,
-        manifest,
-    })
-}
-
-pub async fn read_snapshot_manifest(
-    blobs: &(impl BlobStore + ?Sized),
-    snapshot_ref: &BlobRef,
-) -> Result<VfsSnapshotManifest, VfsError> {
-    let bytes = blobs.read_bytes(snapshot_ref).await?;
-    decode_snapshot_manifest(&bytes)
-}
-
-pub fn encode_snapshot_manifest(manifest: &VfsSnapshotManifest) -> Result<Vec<u8>, VfsError> {
-    manifest.validate()?;
-    serde_json::to_vec(manifest).map_err(|source| VfsError::EncodeManifest { source })
-}
-
-pub fn decode_snapshot_manifest(bytes: &[u8]) -> Result<VfsSnapshotManifest, VfsError> {
-    let manifest: VfsSnapshotManifest =
-        serde_json::from_slice(bytes).map_err(|source| VfsError::DecodeManifest { source })?;
-    manifest.validate()?;
-    Ok(manifest)
-}
-
-pub fn lookup_snapshot_path<'a>(
-    manifest: &'a VfsSnapshotManifest,
-    path: &VfsPath,
-) -> Result<VfsNode<'a>, VfsError> {
-    if path.is_root() {
-        return Ok(VfsNode::Directory(&manifest.root));
-    }
-
-    let mut current = &manifest.root;
-    let components = path.components();
-    for (index, component) in components.iter().enumerate() {
-        let entry = current
-            .entries
-            .get(*component)
-            .ok_or_else(|| VfsError::NotFound { path: path.clone() })?;
-        let is_last = index == components.len() - 1;
-        match (entry, is_last) {
-            (VfsEntry::File(file), true) => return Ok(VfsNode::File(file)),
-            (VfsEntry::Directory(directory), true) => return Ok(VfsNode::Directory(directory)),
-            (VfsEntry::Directory(directory), false) => current = directory,
-            (VfsEntry::File(_), false) => {
-                return Err(VfsError::NotADirectory { path: path.clone() });
-            }
-        }
-    }
-
-    Err(VfsError::NotFound { path: path.clone() })
-}
-
-pub async fn read_snapshot_file(
-    blobs: &(impl BlobStore + ?Sized),
-    manifest: &VfsSnapshotManifest,
-    path: &VfsPath,
-) -> Result<Vec<u8>, VfsError> {
-    let file = match lookup_snapshot_path(manifest, path)? {
-        VfsNode::File(file) => file,
-        VfsNode::Directory(_) => return Err(VfsError::NotAFile { path: path.clone() }),
-    };
-    let bytes = blobs.read_bytes(&file.blob_ref).await?;
-    if bytes.len() as u64 != file.size_bytes {
-        return Err(VfsError::InvalidManifest {
-            message: format!(
-                "file size mismatch for {path}: manifest has {}, blob has {}",
-                file.size_bytes,
-                bytes.len()
-            ),
-        });
-    }
-    Ok(bytes)
-}
-
-pub fn list_snapshot_directory(
-    manifest: &VfsSnapshotManifest,
-    path: &VfsPath,
-) -> Result<Vec<VfsDirectoryListingEntry>, VfsError> {
-    let directory = match lookup_snapshot_path(manifest, path)? {
-        VfsNode::Directory(directory) => directory,
-        VfsNode::File(_) => return Err(VfsError::NotADirectory { path: path.clone() }),
-    };
-
-    Ok(directory
-        .entries
-        .iter()
-        .map(|(file_name, entry)| match entry {
-            VfsEntry::File(file) => VfsDirectoryListingEntry {
-                file_name: file_name.clone(),
-                is_directory: false,
-                is_file: true,
-                size_bytes: Some(file.size_bytes),
-            },
-            VfsEntry::Directory(_) => VfsDirectoryListingEntry {
-                file_name: file_name.clone(),
-                is_directory: true,
-                is_file: false,
-                size_bytes: None,
-            },
-        })
-        .collect())
-}
-
-pub fn stat_snapshot_path(
-    manifest: &VfsSnapshotManifest,
-    path: &VfsPath,
-) -> Result<VfsPathMetadata, VfsError> {
-    match lookup_snapshot_path(manifest, path)? {
-        VfsNode::File(file) => Ok(VfsPathMetadata {
-            is_directory: false,
-            is_file: true,
-            size_bytes: Some(file.size_bytes),
-            media_type: file.media_type.clone(),
-            executable: file.executable,
-        }),
-        VfsNode::Directory(_) => Ok(VfsPathMetadata {
-            is_directory: true,
-            is_file: false,
-            size_bytes: None,
-            media_type: None,
-            executable: false,
-        }),
-    }
-}
-
-fn validate_inline_files(files: &[InlineFile], limits: VfsSnapshotLimits) -> Result<(), VfsError> {
-    if files.len() as u64 > limits.max_files {
-        return Err(VfsError::LimitExceeded {
-            limit: "files",
-            value: files.len() as u64,
-            max: limits.max_files,
-        });
-    }
-
-    let mut total_bytes = 0u64;
-    let mut seen = BTreeSet::new();
-    for file in files {
-        if file.path.is_root() {
-            return Err(VfsError::RootFile);
-        }
-        if !seen.insert(file.path.clone()) {
-            return Err(VfsError::DuplicatePath {
-                path: file.path.clone(),
-            });
-        }
-        if file.path.depth() > limits.max_depth {
-            return Err(VfsError::LimitExceeded {
-                limit: "depth",
-                value: file.path.depth() as u64,
-                max: limits.max_depth as u64,
-            });
-        }
-        let file_bytes = file.bytes.len() as u64;
-        if file_bytes > limits.max_single_file_bytes {
-            return Err(VfsError::LimitExceeded {
-                limit: "single_file_bytes",
-                value: file_bytes,
-                max: limits.max_single_file_bytes,
-            });
-        }
-        total_bytes = total_bytes
-            .checked_add(file_bytes)
-            .ok_or(VfsError::LimitExceeded {
-                limit: "total_bytes",
-                value: u64::MAX,
-                max: limits.max_total_bytes,
-            })?;
-        if total_bytes > limits.max_total_bytes {
-            return Err(VfsError::LimitExceeded {
-                limit: "total_bytes",
-                value: total_bytes,
-                max: limits.max_total_bytes,
-            });
-        }
-    }
-
-    let mut root = VfsDirectory::default();
-    for file in files {
-        insert_file(
-            &mut root,
-            &file.path,
-            VfsFile {
-                blob_ref: BlobRef::default(),
-                size_bytes: file.bytes.len() as u64,
-                media_type: file.media_type.clone(),
-                executable: file.executable,
-            },
-        )?;
-    }
-
-    Ok(())
-}
-
-fn insert_file(
-    directory: &mut VfsDirectory,
-    path: &VfsPath,
-    file: VfsFile,
-) -> Result<(), VfsError> {
-    let components = path.components();
-    if components.is_empty() {
-        return Err(VfsError::RootFile);
-    }
-
-    let mut current = directory;
-    for component in &components[..components.len() - 1] {
-        let entry = current
-            .entries
-            .entry((*component).to_owned())
-            .or_insert_with(|| VfsEntry::Directory(VfsDirectory::default()));
-        match entry {
-            VfsEntry::Directory(directory) => {
-                current = directory;
-            }
-            VfsEntry::File(_) => {
-                return Err(VfsError::PathConflict {
-                    path: path.clone(),
-                    existing: "file",
-                });
-            }
-        }
-    }
-
-    let filename = components[components.len() - 1].to_owned();
-    match current.entries.insert(filename, VfsEntry::File(file)) {
-        None => Ok(()),
-        Some(VfsEntry::Directory(_)) => Err(VfsError::PathConflict {
-            path: path.clone(),
-            existing: "directory",
-        }),
-        Some(VfsEntry::File(_)) => Err(VfsError::DuplicatePath { path: path.clone() }),
-    }
-}
-
-fn validate_directory(directory: &VfsDirectory, totals: &mut VfsTotals) -> Result<(), VfsError> {
-    for (name, entry) in &directory.entries {
-        validate_name(name)?;
-        match entry {
-            VfsEntry::File(file) => {
-                totals.files += 1;
-                totals.bytes =
-                    totals
-                        .bytes
-                        .checked_add(file.size_bytes)
-                        .ok_or(VfsError::InvalidManifest {
-                            message: "file byte total overflowed u64".to_owned(),
-                        })?;
-            }
-            VfsEntry::Directory(directory) => validate_directory(directory, totals)?,
-        }
-    }
-    Ok(())
-}
-
-fn validate_name(name: &str) -> Result<(), VfsError> {
-    if name.is_empty() || name == "." || name == ".." || name.as_bytes().contains(&0) {
-        return Err(VfsError::InvalidManifest {
-            message: format!("invalid directory entry name '{name}'"),
-        });
-    }
-    if name.contains('/') {
-        return Err(VfsError::InvalidManifest {
-            message: format!("directory entry name contains '/': '{name}'"),
-        });
-    }
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use engine::storage::InMemoryBlobStore;
-
-    #[test]
-    fn vfs_path_normalizes_relative_and_absolute_paths() {
-        assert_eq!(VfsPath::parse("foo/bar").unwrap().as_str(), "/foo/bar");
-        assert_eq!(VfsPath::parse("/foo/bar").unwrap().as_str(), "/foo/bar");
-        assert_eq!(VfsPath::parse("/").unwrap(), VfsPath::root());
-        assert_eq!(
-            VfsPath::parse("foo bar/baz.txt").unwrap().as_str(),
-            "/foo bar/baz.txt"
-        );
-    }
-
-    #[test]
-    fn vfs_path_rejects_unsafe_components() {
-        assert!(matches!(VfsPath::parse(""), Err(VfsPathError::Empty)));
-        assert!(matches!(
-            VfsPath::parse("foo//bar"),
-            Err(VfsPathError::InvalidComponent { .. })
-        ));
-        assert!(matches!(
-            VfsPath::parse("foo/./bar"),
-            Err(VfsPathError::InvalidComponent { .. })
-        ));
-        assert!(matches!(
-            VfsPath::parse("foo/../bar"),
-            Err(VfsPathError::InvalidComponent { .. })
-        ));
-        assert!(matches!(
-            VfsPath::parse("foo/\0/bar"),
-            Err(VfsPathError::ContainsNul { .. })
-        ));
-    }
-
-    #[test]
-    fn vfs_workspace_id_uses_engine_string_id_rules() {
-        let id = VfsWorkspaceId::new("workspace-1");
-
-        assert_eq!(id.as_str(), "workspace-1");
-        assert!(VfsWorkspaceId::try_new("-workspace").is_err());
-        assert!(VfsWorkspaceId::try_new("workspace/path").is_err());
-        assert!(VfsWorkspaceId::try_new("").is_err());
-    }
-
-    #[test]
-    fn catalog_records_round_trip_as_json() {
-        let snapshot_ref = BlobRef::from_bytes(b"manifest");
-        let workspace_id = VfsWorkspaceId::new("workspace-1");
-
-        let snapshot = VfsSnapshotRecord {
-            snapshot_ref: snapshot_ref.clone(),
-            source: VfsSnapshotSource::new("skill")
-                .with_subject("openai-docs")
-                .with_metadata_ref(BlobRef::from_bytes(b"skill metadata")),
-            display_name: Some("OpenAI Docs".to_string()),
-            created_at_ms: 10,
-        };
-        let encoded = serde_json::to_string(&snapshot).unwrap();
-        let decoded: VfsSnapshotRecord = serde_json::from_str(&encoded).unwrap();
-        assert_eq!(decoded, snapshot);
-
-        let workspace = VfsWorkspaceRecord {
-            workspace_id: workspace_id.clone(),
-            base_snapshot_ref: Some(snapshot_ref.clone()),
-            head_snapshot_ref: BlobRef::from_bytes(b"head"),
-            revision: 7,
-            created_at_ms: 11,
-            updated_at_ms: 12,
-        };
-        let encoded = serde_json::to_string(&workspace).unwrap();
-        let decoded: VfsWorkspaceRecord = serde_json::from_str(&encoded).unwrap();
-        assert_eq!(decoded, workspace);
-
-        let mount = VfsMountRecord {
-            session_id: SessionId::new("session-1"),
-            mount_path: VfsPath::parse("/workspace").unwrap(),
-            source: VfsMountSource::Workspace { workspace_id },
-            access: VfsMountAccess::ReadWrite,
-        };
-        assert!(mount.access.is_writable());
-        let encoded = serde_json::to_string(&mount).unwrap();
-        let decoded: VfsMountRecord = serde_json::from_str(&encoded).unwrap();
-        assert_eq!(decoded, mount);
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn inline_snapshot_writes_file_blobs_and_manifest() {
-        let blobs = InMemoryBlobStore::new();
-        let request = CreateInlineSnapshotRequest::new(vec![
-            InlineFile::new("README.md", b"# hello\n".to_vec())
-                .unwrap()
-                .with_media_type("text/markdown"),
-            InlineFile::new("src/lib.rs", b"pub fn ok() {}\n".to_vec()).unwrap(),
-            InlineFile::new("scripts/build.sh", b"#!/bin/sh\n".to_vec())
-                .unwrap()
-                .executable(true),
-        ]);
-
-        let result = create_inline_snapshot(&blobs, request).await.unwrap();
-        assert_eq!(result.manifest.schema_version, VFS_SNAPSHOT_SCHEMA_VERSION);
-        assert_eq!(
-            result.manifest.totals,
-            VfsTotals {
-                files: 3,
-                bytes: 33
-            }
-        );
-
-        let manifest_bytes = blobs.read_bytes(&result.snapshot_ref).await.unwrap();
-        let decoded = decode_snapshot_manifest(&manifest_bytes).unwrap();
-        assert_eq!(decoded, result.manifest);
-
-        let readme = match decoded.root.entry("README.md").unwrap() {
-            VfsEntry::File(file) => file,
-            VfsEntry::Directory(_) => panic!("README.md should be a file"),
-        };
-        assert_eq!(readme.media_type.as_deref(), Some("text/markdown"));
-        assert_eq!(readme.size_bytes, 8);
-        assert_eq!(
-            blobs.read_bytes(&readme.blob_ref).await.unwrap(),
-            b"# hello\n"
-        );
-
-        let scripts = match decoded.root.entry("scripts").unwrap() {
-            VfsEntry::Directory(directory) => directory,
-            VfsEntry::File(_) => panic!("scripts should be a directory"),
-        };
-        let build = match scripts.entry("build.sh").unwrap() {
-            VfsEntry::File(file) => file,
-            VfsEntry::Directory(_) => panic!("build.sh should be a file"),
-        };
-        assert!(build.executable);
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn snapshot_read_helpers_resolve_paths() {
-        let blobs = InMemoryBlobStore::new();
-        let result = create_inline_snapshot(
-            &blobs,
-            CreateInlineSnapshotRequest::new(vec![
-                InlineFile::new("README.md", b"hello\n".to_vec()).unwrap(),
-                InlineFile::new("src/lib.rs", b"pub fn f() {}\n".to_vec())
-                    .unwrap()
-                    .with_media_type("text/rust"),
-            ]),
-        )
-        .await
-        .unwrap();
-
-        let loaded = read_snapshot_manifest(&blobs, &result.snapshot_ref)
-            .await
-            .unwrap();
-        assert_eq!(loaded, result.manifest);
-
-        assert!(matches!(
-            lookup_snapshot_path(&loaded, &VfsPath::root()).unwrap(),
-            VfsNode::Directory(_)
-        ));
-        assert_eq!(
-            read_snapshot_file(&blobs, &loaded, &VfsPath::parse("/README.md").unwrap())
-                .await
-                .unwrap(),
-            b"hello\n"
-        );
-
-        let root_entries = list_snapshot_directory(&loaded, &VfsPath::root()).unwrap();
-        assert_eq!(
-            root_entries,
-            vec![
-                VfsDirectoryListingEntry {
-                    file_name: "README.md".to_string(),
-                    is_directory: false,
-                    is_file: true,
-                    size_bytes: Some(6),
-                },
-                VfsDirectoryListingEntry {
-                    file_name: "src".to_string(),
-                    is_directory: true,
-                    is_file: false,
-                    size_bytes: None,
-                },
-            ]
-        );
-
-        let metadata = stat_snapshot_path(&loaded, &VfsPath::parse("src/lib.rs").unwrap()).unwrap();
-        assert!(metadata.is_file);
-        assert!(!metadata.is_directory);
-        assert_eq!(metadata.size_bytes, Some(14));
-        assert_eq!(metadata.media_type.as_deref(), Some("text/rust"));
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn snapshot_read_helpers_report_missing_and_wrong_kind_paths() {
-        let blobs = InMemoryBlobStore::new();
-        let result = create_inline_snapshot(
-            &blobs,
-            CreateInlineSnapshotRequest::new(vec![
-                InlineFile::new("README.md", b"hello\n".to_vec()).unwrap(),
-                InlineFile::new("src/lib.rs", b"pub fn f() {}\n".to_vec()).unwrap(),
-            ]),
-        )
-        .await
-        .unwrap();
-
-        assert!(matches!(
-            lookup_snapshot_path(&result.manifest, &VfsPath::parse("/missing").unwrap()),
-            Err(VfsError::NotFound { .. })
-        ));
-        assert!(matches!(
-            read_snapshot_file(&blobs, &result.manifest, &VfsPath::parse("/src").unwrap()).await,
-            Err(VfsError::NotAFile { .. })
-        ));
-        assert!(matches!(
-            list_snapshot_directory(&result.manifest, &VfsPath::parse("/README.md").unwrap()),
-            Err(VfsError::NotADirectory { .. })
-        ));
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn inline_snapshot_allows_empty_tree() {
-        let blobs = InMemoryBlobStore::new();
-        let result = create_inline_snapshot(&blobs, CreateInlineSnapshotRequest::new(Vec::new()))
-            .await
-            .unwrap();
-
-        assert_eq!(result.manifest, VfsSnapshotManifest::empty());
-        assert!(blobs.has_blob(&result.snapshot_ref).await.unwrap());
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn inline_snapshot_rejects_duplicate_paths_before_writing() {
-        let blobs = InMemoryBlobStore::new();
-        let request = CreateInlineSnapshotRequest::new(vec![
-            InlineFile::new("a.txt", b"one".to_vec()).unwrap(),
-            InlineFile::new("/a.txt", b"two".to_vec()).unwrap(),
-        ]);
-
-        let error = create_inline_snapshot(&blobs, request).await.unwrap_err();
-        assert!(matches!(error, VfsError::DuplicatePath { .. }));
-        assert!(!blobs.has_blob(&BlobRef::from_bytes(b"one")).await.unwrap());
-        assert!(!blobs.has_blob(&BlobRef::from_bytes(b"two")).await.unwrap());
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn inline_snapshot_rejects_file_directory_conflicts_before_writing() {
-        let blobs = InMemoryBlobStore::new();
-        let request = CreateInlineSnapshotRequest::new(vec![
-            InlineFile::new("src", b"file".to_vec()).unwrap(),
-            InlineFile::new("src/lib.rs", b"nested".to_vec()).unwrap(),
-        ]);
-
-        let error = create_inline_snapshot(&blobs, request).await.unwrap_err();
-        assert!(matches!(
-            error,
-            VfsError::PathConflict {
-                existing: "file",
-                ..
-            }
-        ));
-        assert!(!blobs.has_blob(&BlobRef::from_bytes(b"file")).await.unwrap());
-        assert!(
-            !blobs
-                .has_blob(&BlobRef::from_bytes(b"nested"))
-                .await
-                .unwrap()
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn inline_snapshot_enforces_limits_before_writing() {
-        let blobs = InMemoryBlobStore::new();
-        let request = CreateInlineSnapshotRequest::new(vec![
-            InlineFile::new("one.txt", b"1234".to_vec()).unwrap(),
-            InlineFile::new("two.txt", b"56".to_vec()).unwrap(),
-        ])
-        .with_limits(VfsSnapshotLimits::new(10, 5, 10, 10));
-
-        let error = create_inline_snapshot(&blobs, request).await.unwrap_err();
-        assert!(matches!(
-            error,
-            VfsError::LimitExceeded {
-                limit: "total_bytes",
-                value: 6,
-                max: 5
-            }
-        ));
-        assert!(!blobs.has_blob(&BlobRef::from_bytes(b"1234")).await.unwrap());
-        assert!(!blobs.has_blob(&BlobRef::from_bytes(b"56")).await.unwrap());
-    }
-
-    #[tokio::test(flavor = "current_thread")]
-    async fn inline_snapshot_rejects_root_file_and_deep_paths() {
-        let blobs = InMemoryBlobStore::new();
-        let root_request =
-            CreateInlineSnapshotRequest::new(vec![InlineFile::new("/", b"root".to_vec()).unwrap()]);
-        assert!(matches!(
-            create_inline_snapshot(&blobs, root_request)
-                .await
-                .unwrap_err(),
-            VfsError::RootFile
-        ));
-
-        let deep_request = CreateInlineSnapshotRequest::new(vec![
-            InlineFile::new("a/b/c", b"deep".to_vec()).unwrap(),
-        ])
-        .with_limits(VfsSnapshotLimits::new(10, 100, 100, 2));
-        assert!(matches!(
-            create_inline_snapshot(&blobs, deep_request)
-                .await
-                .unwrap_err(),
-            VfsError::LimitExceeded { limit: "depth", .. }
-        ));
-    }
-
-    #[test]
-    fn manifest_encode_decode_validates_schema_and_totals() {
-        let mut manifest = VfsSnapshotManifest::empty();
-        manifest.schema_version = "wrong".to_owned();
-        assert!(matches!(
-            encode_snapshot_manifest(&manifest),
-            Err(VfsError::InvalidManifest { .. })
-        ));
-
-        let mut manifest = VfsSnapshotManifest::empty();
-        manifest.root.entries.insert(
-            "file.txt".to_owned(),
-            VfsEntry::File(VfsFile {
-                blob_ref: BlobRef::from_bytes(b"file"),
-                size_bytes: 4,
-                media_type: None,
-                executable: false,
-            }),
-        );
-        assert!(matches!(
-            encode_snapshot_manifest(&manifest),
-            Err(VfsError::InvalidManifest { .. })
-        ));
-    }
-}
+pub mod catalog;
+pub mod manifest;
+pub mod path;
+pub mod snapshot;
+
+pub use catalog::*;
+pub use manifest::*;
+pub use path::*;
+pub use snapshot::*;
